[FIX] cores em dailyScreen.
